+++ signer-service-rust/.env.example
@@ -1,0 +1,13 @@
+SERVER_HOST=127.0.0.1
+SERVER_PORT=3001
+
+DATABASE_HOST=localhost
+DATABASE_PORT=5432
+POSTGRES_USER=superuser
+POSTGRES_PASSWORD=1234
+DATABASE_URL="postgres://superuser:1234@localhost:5432/postgres"
+
+STELLAR_SECRET_KEY=SCVJD7BHU5LNFXNIDC7E226HISKUOZUEPJWLA2YU2GNBFMP5PYF2TQBH
+STELLAR_PUBLIC_NETWORK=false
+
+
