+++ signer-service-rust/src/api/routes/mod.rs
@@ -1,24 +1,30 @@
 
 /// implementations of the routes
+mod stellar_impls;
 mod impls;
 
 use axum::Router;
 use axum::routing::{get, post};
-use crate::api::routes::impls::{create_account, payment, status};
+use crate::api::routes::impls::tokens;
+use crate::api::routes::stellar_impls::{create_account, payment, status};
 use crate::AppState;
 
 /// Returns /v1/... routes
 pub fn v1_routes(state:AppState) -> Router {
     Router::new().nest(
         "/v1",
-        status_route(state.clone())
+        first_level(state.clone())
             .merge(stellar_routes(state))
     )
 }
 
-/// GET /v1/status
-fn status_route(state:AppState) -> Router {
-    Router::new().route("/status", get(status))
+/// contains the /status and /tokens
+fn first_level(state:AppState) -> Router {
+    Router::new()
+        /// GET /v1/status
+        .route("/status", get(status))
+        /// GET /v1/tokens
+        .route("/tokens", get(tokens))
         .with_state(state)
 }
 
@@ -29,5 +35,5 @@ fn stellar_routes(state:AppState) -> Router {
         .route("/stellar/create", post(create_account))
         .route("/stellar/payment", post(payment))
         .with_state(state)
-
 }
+
